gitignore & bootloader
pacman.cof was added to .gitignore
We don't need it to be pushed as the build.sh script is made to render
from a template our pacman.conf.
Both systemd-boot and syslinux first entry was modified to include
cow_spacesize=2G to increase the CoW space when running the system.
Next release may have it considered a template too with custom boot
options